+++ crd-generator/api/src/main/java/io/fabric8/crd/generator/AbstractJsonSchema.java
@@ -49,10 +49,12 @@ public abstract class AbstractJsonSchema<T, B> {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(AbstractJsonSchema.class);
 
+  protected static final TypeDef OBJECT = TypeDef.forName(Object.class.getName());
   protected static final TypeDef QUANTITY = TypeDef.forName(Quantity.class.getName());
   protected static final TypeDef DURATION = TypeDef.forName(Duration.class.getName());
-  protected static final TypeDef INT_OR_STRING =TypeDef.forName(IntOrString.class.getName());  
+  protected static final TypeDef INT_OR_STRING =TypeDef.forName(IntOrString.class.getName());
 
+  protected static final TypeRef OBJECT_REF = OBJECT.toReference();
   protected static final TypeRef QUANTITY_REF = QUANTITY.toReference();
   protected static final TypeRef DURATION_REF = DURATION.toReference();
   protected static final TypeRef INT_OR_STRING_REF = INT_OR_STRING.toReference();
@@ -76,7 +78,10 @@ public abstract class AbstractJsonSchema<T, B> {
   public static final String ANNOTATION_JSON_PROPERTY = "com.fasterxml.jackson.annotation.JsonProperty";
   public static final String ANNOTATION_JSON_PROPERTY_DESCRIPTION = "com.fasterxml.jackson.annotation.JsonPropertyDescription";
   public static final String ANNOTATION_JSON_IGNORE = "com.fasterxml.jackson.annotation.JsonIgnore";
+  public static final String ANNOTATION_JSON_ANY_GETTER = "com.fasterxml.jackson.annotation.JsonAnyGetter";
+  public static final String ANNOTATION_JSON_ANY_SETTER = "com.fasterxml.jackson.annotation.JsonAnySetter";
   public static final String ANNOTATION_NOT_NULL = "javax.validation.constraints.NotNull";
+  public static final String ANNOTATION_SCHEMA_FROM = "io.fabric8.crd.generator.annotation.SchemaFrom";
 
   public static final String JSON_NODE_TYPE = "com.fasterxml.jackson.databind.JsonNode";
 
@@ -125,7 +130,7 @@ public abstract class AbstractJsonSchema<T, B> {
         .emptySet();
     List<String> required = new ArrayList<>();
 
-    final boolean preserveUnknownFields = (
+    boolean preserveUnknownFields = (
       definition.getFullyQualifiedName() != null &&
         definition.getFullyQualifiedName().equals(JSON_NODE_TYPE));
 
@@ -149,6 +154,10 @@ public abstract class AbstractJsonSchema<T, B> {
         continue;
       }
       final T schema = internalFromImpl(name, possiblyRenamedProperty.getTypeRef(), visited);
+      if (facade.preserveUnknownFields) {
+        preserveUnknownFields = true;
+      }
+
       // if we got a description from the field or an accessor, use it
       final String description = facade.description;
       final T possiblyUpdatedSchema;
@@ -180,7 +189,9 @@ public abstract class AbstractJsonSchema<T, B> {
     private String renamedTo;
     private boolean required;
     private boolean ignored;
+    private boolean preserveUnknownFields;
     private String description;
+    private Class schemaFrom;
 
     private PropertyOrAccessor(Collection<AnnotationRef> annotations, String name, String propertyName, boolean isMethod) {
       this.annotations = annotations;
@@ -217,6 +228,15 @@ public abstract class AbstractJsonSchema<T, B> {
             break;
           case ANNOTATION_JSON_IGNORE:
             ignored = true;
+          case ANNOTATION_JSON_ANY_GETTER:
+          case ANNOTATION_JSON_ANY_SETTER:
+            preserveUnknownFields = true;
+            break;
+          case ANNOTATION_SCHEMA_FROM:
+            final Class extractedType = (Class) a.getParameters().get("type");
+            if (extractedType != null) {
+              schemaFrom = extractedType;
+            }
             break;
         }
       });
@@ -234,6 +254,10 @@ public abstract class AbstractJsonSchema<T, B> {
       return ignored;
     }
 
+    public boolean isPreserveUnknownFields() {
+      return preserveUnknownFields;
+    }
+
     public String getDescription() {
       return description;
     }
@@ -244,6 +268,14 @@ public abstract class AbstractJsonSchema<T, B> {
 
     public boolean contributeDescription() {
       return description != null;
+    }
+
+    public Class getSchemaFrom() {
+      return schemaFrom;
+    }
+
+    public boolean contributeSchemaFrom() {
+      return schemaFrom != null;
     }
 
     @Override
@@ -258,9 +290,11 @@ public abstract class AbstractJsonSchema<T, B> {
     private String description;
     private boolean required;
     private boolean ignored;
+    private boolean preserveUnknownFields;
     private final Property original;
     private String nameContributedBy;
     private String descriptionContributedBy;
+    private TypeRef schemaFrom;
 
     public PropertyFacade(Property property, Map<String, Method> potentialAccessors) {
       original = property;
@@ -310,10 +344,21 @@ public abstract class AbstractJsonSchema<T, B> {
         } else if (p.isIgnored()) {
           ignored = true;
         }
+
+        if (p.isPreserveUnknownFields()) {
+          preserveUnknownFields = true;
+        }
+
+        if (p.contributeSchemaFrom()) {
+          schemaFrom = Types.typeDefFrom(p.getSchemaFrom()).toReference();
+        }
       });
-      
-      return renamedTo != null ? new Property(original.getAnnotations(), original.getTypeRef(), renamedTo,
-        original.getComments(), original.getModifiers(), original.getAttributes()) : original;
+
+      TypeRef typeRef = schemaFrom != null ? schemaFrom : original.getTypeRef();
+      String finalName = renamedTo != null ? renamedTo : original.getName();
+
+      return new Property(original.getAnnotations(), typeRef, finalName,
+        original.getComments(), original.getModifiers(), original.getAttributes());
     }
   }
 
@@ -393,25 +438,20 @@ public abstract class AbstractJsonSchema<T, B> {
       return arrayLikeProperty(schema);
     } else if (io.sundr.model.utils.Collections.IS_MAP.apply(typeRef)) { // Handle Maps
       final TypeRef keyType = TypeAs.UNWRAP_MAP_KEY_OF.apply(typeRef);
-      boolean degraded = false;
-      if (keyType instanceof ClassRef) {
-        ClassRef classRef = (ClassRef) keyType;
-        if (!classRef.getFullyQualifiedName().equals("java.lang.String")) {
-          degraded = true;
-        } else {
-          final TypeRef valueType = TypeAs.UNWRAP_MAP_VALUE_OF.apply(typeRef);
-          classRef = (ClassRef) valueType;
-          if (!classRef.getFullyQualifiedName().equals("java.lang.String")) {
-            degraded = true;
-          }
-        }
-      } else {
-        degraded = true;
+
+      if (!(keyType instanceof ClassRef && ((ClassRef) keyType).getFullyQualifiedName().equals("java.lang.String"))) {
+        LOGGER.warn("Property '{}' with '{}' key type is mapped to 'string' because of CRD schemas limitations", name, typeRef);
       }
-      if (degraded) {
-        LOGGER.warn("Property '{}' with '{}' type is mapped to a string to string mapping because of CRD schemas limitations", name, typeRef);
+
+      final TypeRef valueType = TypeAs.UNWRAP_MAP_VALUE_OF.apply(typeRef);
+      T schema = internalFromImpl(name, valueType, visited);
+
+      if (schema == null) {
+        LOGGER.warn("Property '{}' with '{}' value type is mapped to 'object' because its CRD representation cannot be extracted.", name, typeRef);
+        schema = internalFromImpl(name, OBJECT_REF, visited);
       }
-      return mapLikeProperty();
+
+      return mapLikeProperty(schema);
     } else if (io.sundr.model.utils.Optionals.isOptional(typeRef)) { // Handle Optionals
       return internalFromImpl(name, TypeAs.UNWRAP_OPTIONAL_OF.apply(typeRef), visited);
     } else {
@@ -484,9 +524,10 @@ public abstract class AbstractJsonSchema<T, B> {
   /**
    * Builds the schema for map-like properties
    *
+   * @param schema the schema for the extracted element type for the values of this map-like property
    * @return the schema for the map-like property
    */
-  protected abstract T mapLikeProperty();
+  protected abstract T mapLikeProperty(T schema);
 
   /**
    * Builds the schema for standard, simple (e.g. string) property types