+++ crd-generator/api/src/test/java/io/fabric8/crd/example/extraction/Extraction.java
@@ -13,24 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd.example.json;
+package io.fabric8.crd.example.extraction;
 
-import com.fasterxml.jackson.databind.JsonNode;
+import io.fabric8.kubernetes.client.CustomResource;
 
-public class ContainingJsonSpec {
-
-  private int field;
-
-  public int getField() { return field; }
-
-  private JsonNode free;
-
-  public JsonNode getFree() {
-    return free;
-  }
-
-  private Foo foo;
-
-  public Foo getFoo() { return foo; }
+public class Extraction extends  CustomResource<ExtractionSpec, Void> {
 
 }