+++ crd-generator/api/src/test/java/io/fabric8/crd/example/map/ContainingMaps.java
@@ -13,24 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd.example.json;
+package io.fabric8.crd.example.map;
 
-import com.fasterxml.jackson.databind.JsonNode;
+import io.fabric8.kubernetes.client.CustomResource;
+import io.fabric8.kubernetes.model.annotation.Group;
+import io.fabric8.kubernetes.model.annotation.Version;
 
-public class ContainingJsonSpec {
-
-  private int field;
-
-  public int getField() { return field; }
-
-  private JsonNode free;
-
-  public JsonNode getFree() {
-    return free;
-  }
-
-  private Foo foo;
-
-  public Foo getFoo() { return foo; }
+@Group("map.fabric8.io")
+@Version("v1alpha1")
+public class ContainingMaps extends  CustomResource<ContainingMapsSpec, Void> {
 
 }